+++ CHANGELOG.md
@@ -22,3 +22,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased] - 2022-MM-DD
+
+### Added
+
+## [Unreleased] - 2022-MM-DD
+
+### Added
+
+- Setup the project structure. ([#385](https://github.com/jina-ai/finetuner/pull/385))
+
+- Create experiment endpoints. ([#386](https://github.com/jina-ai/finetuner/pull/386))
+
+### Removed
+
+- Delete all unnecessary files from the previous project. ([#384](https://github.com/jina-ai/finetuner/pull/384))